[quant][refactor tests] Rename test_numeric_suite and equalization tests
Summary:
Rename the test files to be more explicit that they are for eager mode

Test Plan:
python test/test_quantization.py

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
