+++ test/test_quantization.py
@@ -72,7 +72,7 @@ except ImportError:
     pass
 
 # Tooling: numeric_suite
-from quantization.test_numeric_suite import TestEagerModeNumericSuite  # noqa: F401
+from quantization.test_numeric_suite_eager import TestEagerModeNumericSuite  # noqa: F401
 
 try:
     from quantization.test_numeric_suite_fx import TestFXGraphMatcher  # noqa: F401
@@ -86,9 +86,9 @@ except ImportError:
 from quantization.test_backward_compatibility import TestSerialization  # noqa: F401
 
 # Equalization
-from quantization.test_equalize import TestEqualizeEager  # noqa: F401
+from quantization.test_equalize_eager import TestEqualizeEager  # noqa: F401
 # Bias Correction
-from quantization.test_bias_correction import TestBiasCorrection  # noqa: F401
+from quantization.test_bias_correction_eager import TestBiasCorrection  # noqa: F401
 
 if __name__ == '__main__':
     run_tests()
